requirements.txt
langchain
langchain_openai
ipykernel
python-dotenv
langchain_community
PyPDF
beautifulsoup4
playwright
arxiv
pymupdf
langchain-text-splitters
python-dotenv
chromadb
llama-index-embeddings-ollama
#ollama
scikit-learn
langchain_huggingface
sentence_transformers
faiss-cpu
chromadb==0.4.22
langchain-chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
pydantic==1.10.13
